+++ src/components/filled-input/filled-input.tsx
@@ -1,0 +1,14 @@
+import React, {FC} from "react";
+import {FormControlProps} from "@mui/material";
+import {StyledFilledInput} from "./styled";
+
+export const FilledInput: FC<FormControlProps> = (props, {
+    ...restProps
+}) => {
+    return (
+        <StyledFilledInput
+            {...props}
+            {...restProps}
+        />
+    )
+}
