+++ src/components/filled-input/index.ts
@@ -1,0 +1,2 @@
+export * from './filled-input';
+export * from './theme';
