+++ src/components/filled-input/styled.ts
@@ -1,0 +1,4 @@
+import {styled} from "@mui/material/styles";
+import {FilledInput as MuiFilledInput} from "@mui/material";
+
+export const StyledFilledInput = styled(MuiFilledInput)({});
