+++ src/components/form-control-label/index.ts
@@ -1,0 +1,2 @@
+export * from './form-control-label';
+export * from './theme';
