+++ src/components/form-control-label/styled.ts
@@ -1,0 +1,4 @@
+import {styled} from "@mui/material/styles";
+import {FormControlLabel as MuiFormControlLabel} from "@mui/material";
+
+export const StyledFormControlLabel = styled(MuiFormControlLabel)({});
