+++ src/components/form-control-label/theme.ts
@@ -1,0 +1,11 @@
+import {Components} from "@mui/material/styles/components";
+
+export const FormControlLabelTheme = (theme:any): Components["MuiFormControlLabel"]=> (
+    {
+        defaultProps: {},
+        styleOverrides: {
+            root: {
+            }
+        }
+    }
+)
