+++ src/components/form-control/form-control.tsx
@@ -1,0 +1,14 @@
+import React, {FC} from "react";
+import {FormControlProps} from "@mui/material";
+import {StyledFormControl} from "./styled";
+
+export const FormControl: FC<FormControlProps> = (props, {
+    ...restProps
+}) => {
+    return (
+        <StyledFormControl
+            {...props}
+            {...restProps}
+        />
+    )
+}
