+++ src/components/form-control/index.ts
@@ -1,0 +1,2 @@
+export * from './form-control';
+export * from './theme';
