+++ src/components/form-control/styled.ts
@@ -1,0 +1,4 @@
+import {styled} from "@mui/material/styles";
+import {FormControl as MuiFormControl} from "@mui/material";
+
+export const StyledFormControl = styled(MuiFormControl)({});
