+++ src/components/form-control/theme.ts
@@ -1,0 +1,11 @@
+import {Components} from "@mui/material/styles/components";
+
+export const FormControlTheme = (theme:any): Components["MuiFormControl"]=> (
+    {
+        defaultProps: {},
+        styleOverrides: {
+            root: {
+            }
+        }
+    }
+)
