+++ src/components/form-group/form-group.tsx
@@ -1,0 +1,14 @@
+import React, {FC} from "react";
+import {FormGroupProps} from "@mui/material";
+import {StyledFormGroup} from "./styled";
+
+export const FormGroup: FC<FormGroupProps> = (props, {
+    ...restProps
+}) => {
+    return (
+        <StyledFormGroup
+            {...props}
+            {...restProps}
+        />
+    )
+}
