+++ src/components/form-group/index.ts
@@ -1,0 +1,2 @@
+export * from './form-group';
+export * from './theme';
