+++ src/components/form-group/styled.ts
@@ -1,0 +1,4 @@
+import {styled} from "@mui/material/styles";
+import {FormGroup as MuiFormGroup} from "@mui/material";
+
+export const StyledFormGroup = styled(MuiFormGroup)({});
