+++ src/components/form-group/theme.ts
@@ -1,0 +1,11 @@
+import {Components} from "@mui/material/styles/components";
+
+export const FormGroupTheme = (theme:any): Components["MuiFormGroup"]=> (
+    {
+        defaultProps: {},
+        styleOverrides: {
+            root: {
+            }
+        }
+    }
+)
