+++ src/components/form-helper-text/form-helper-text.tsx
@@ -1,0 +1,14 @@
+import React, {FC} from "react";
+import {FormHelperTextProps} from "@mui/material";
+import {StyledFormHelperText} from "./styled";
+
+export const FormHelperText: FC<FormHelperTextProps> = (props, {
+    ...restProps
+}) => {
+    return (
+        <StyledFormHelperText
+            {...props}
+            {...restProps}
+        />
+    )
+}
