+++ src/components/form-helper-text/index.ts
@@ -1,0 +1,2 @@
+export * from './form-helper-text';
+export * from './theme';
