+++ src/components/form-helper-text/styled.ts
@@ -1,0 +1,4 @@
+import {styled} from "@mui/material/styles";
+import {FormHelperText as MuiFormHelperText} from "@mui/material";
+
+export const StyledFormHelperText = styled(MuiFormHelperText)({});
