+++ src/components/form-label/form-label.tsx
@@ -1,0 +1,14 @@
+import React, {FC} from "react";
+import {FormLabelProps} from "@mui/material";
+import {StyledFormLabel} from "./styled";
+
+export const FormLabel: FC<FormLabelProps> = (props, {
+    ...restProps
+}) => {
+    return (
+        <StyledFormLabel
+            {...props}
+            {...restProps}
+        />
+    )
+}
