+++ src/components/form-label/index.ts
@@ -1,0 +1,2 @@
+export * from './form-label';
+export * from './theme';
