+++ src/components/form-label/styled.ts
@@ -1,0 +1,4 @@
+import {styled} from "@mui/material/styles";
+import {FormLabel as MuiFormLabel} from "@mui/material";
+
+export const StyledFormLabel = styled(MuiFormLabel)({});
