+++ src/components/form-label/theme.ts
@@ -1,0 +1,12 @@
+import {Components} from "@mui/material/styles/components";
+
+export const FormLabelTheme = (theme:any): Components["MuiFormLabel"]=> (
+    {
+        defaultProps: {},
+        styleOverrides: {
+            root: {
+                fontSize: `0.875rem`,
+            }
+        }
+    }
+)
