+++ src/components/index.ts
@@ -3,3 +3,10 @@ export * from './button';
 export * from './icon-button';
 export * from './chip';
 export * from './text-field';
+export * from './form-control';
+export * from './form-control-label';
+export * from './form-group';
+export * from './form-helper-text';
+export * from './form-label';
+export * from './outlined-input';
+export * from './filled-input';
