+++ src/components/outlined-input/index.ts
@@ -1,0 +1,2 @@
+export * from './outlined-input';
+export * from './theme';
