+++ src/components/outlined-input/styled.ts
@@ -1,0 +1,4 @@
+import {styled} from "@mui/material/styles";
+import {OutlinedInput as MuiOutlinedInput} from "@mui/material";
+
+export const StyledOutlinedInput = styled(MuiOutlinedInput)({});
